requirements.txt
flask==3.0.0
werkzeug==3.0.1
google-generativeai>=0.4.0
python-dotenv==0.19.0